Add eventListeners to checkboxes, updating completed status true/false
