+++ src/item.js
@@ -1,0 +1,8 @@
+/* eslint-disable import/prefer-default-export */
+
+export class Item {
+  constructor(text) {
+    this.text = text;
+    this.completed = false;
+  }
+}
